requirements.txt
bcrypt==3.1.0
cffi==1.7.0
click==6.6
Flask==0.11.1
Flask-Cors==3.0.2
Flask-Migrate==2.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-Testing==0.5.0
Flask-graphql
graphene==1.4.0
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
mysql-connector==2.1.4
matplotlib==2.0.2
numpy==1.12.1
py==1.4.31
py-bcrypt==0.4
pycparser==2.14
pygame==1.9.3
pymunk==5.2.0
pytest==3.0.1
pytest-cov==2.3.1
pytest-flask==0.10.0
python-editor==1.0.1
requests==2.14.2
six==1.10.0
SQLAlchemy==1.0.14
Tensorflow>=1.0.0
tflearn-0.3.1
Werkzeug==0.11.10
